Convert dojo/ modules and test files to use has("xyz") rather than dojo.isXYZ (ex: dojo.isIE --> has("ie")).   Also changing dojo/ modules to require dojo/sniff instead of dojo/_base/sniff.  Didn't change modules and test files that aren't yet in AMD format.
Previously, modules we referencing dojo.isQuirks rather than calling has("quirks") because quirks mode can change over time, specifically when a dojo.withDoc() call points dojo to an iframe in a different quirks mode from the main document.   But in that case, the value of has("quirks") also changes, so I think it's OK to use it.

Refs #14614 !strict.

git-svn-id: http://svn.dojotoolkit.org/src/dojo/trunk@27550 560b804f-0ae3-0310-86f3-f6aa0a117693