Modified version of Kitson's patch to support auto require in the parser.
Parser runs the same as before, unless scan() sees undefined modules, in which case it loads the modules asynchronously, and then runs patch-up code on list[], to fill in the missing constructors and remove any nodes that shouldn't be instantiated to due to a stopParser:true setting on an ancestor.

One API change is that scan() now returns a Promise.  That's OK since scan() is new for 1.8.

Fixes #14591 !strict.


git-svn-id: http://svn.dojotoolkit.org/src/dojo/trunk@28469 560b804f-0ae3-0310-86f3-f6aa0a117693