API doc cleanup for dojo/declare.
In order to document methods like inherited() that are available on classes created via dojo/declare, specify the return type of dojo/declare as a class with those methods.   The previous doc comments documented inherited() etc. as methods on Object, which seems both inaccurate (since not all Objects have an inherited() method), and unsuitable for an AMD world, since Object is not an AMD module.

Also inlined the comments for safeMixin() etc.

Refs #14980 !strict


git-svn-id: http://svn.dojotoolkit.org/src/dojo/trunk@28513 560b804f-0ae3-0310-86f3-f6aa0a117693