Slight change in dojo/request API. refs #15182 !strict
dojo/request calls now return a promise that resolves to the handled
data (similar to the old IO calls). The promise also has a "response"
property; this new new property is a promise that resolves to the object
representing the response and can be used to get information like status
and headers (where applicable).

git-svn-id: http://svn.dojotoolkit.org/src/dojo/trunk@28741 560b804f-0ae3-0310-86f3-f6aa0a117693