rpc Intern tests: Replace JSONP test with one using a local service
The Yahoo service used by the existing DOH test appears to be dead.

Unfortunately, this currently doesn't work, because JsonpService uses
dojo/io/script which fails to account for non-variable-friendly characters
such as the hyphen in `dojo-testing`, so we either need to tweak
dojo/io/script or rename `dojo-testing`.

Also moved rpc tests down to browser-only section of all.js.