Correct reading of browser locale
Correct the reading of the browser locale, to avoid situations where the language is taken from the locale from which the browser downloaded, rather than the user's chosen language settings. A common example is where Chrome returns en-US for users in the UK where the navigator.languages[0] would actually return en-GB instead.