Cleanup around getIeDocumentElementOffset().
Remove the calls to getIeDocumentElementOffset() since it's a no-op except for IE6-IE7,
which we desupported in Dojo 1.9.  Also remove the IE6-IE7 code path from getIeDocumentElementOffset(),
but leave the method itself for backwards compatibility.

Finally, use "'pageX' in event" rather than has("dom-addeventlistener") to check if
an event defines the pageX and pageY properties.

Refs #18717.