Resolve #18871, allowing Firefox to load Dojo in iframe with display:none
Firefox returns null from `defaultView.getComputedStyle` in an iframe
with a style of `display: none;`. There is more information in the
discussion in [this thread](https://bugzilla.mozilla.org/show_bug.cgi?id=548397).

NOTE: To ensure the added test ran in Firefox, I had to `npm install
leadfoot@1.7.0` and delete
`node_modules/intern-geezer/node_modules/leadfoot`. This is due to a
need to update the `intern-geezer` library's dependance on an older
version of leadfoot that breaks Firefox 49+ compatibility.