i18n: improve functionality when dojo-v1x-i18n-Api is false (#379)
This change adds a simplified implementation of i18n.getLocalization when
dojo-v1x-i18n-Api is false that allows for some i18n functionality to work.
Without this patch dijit/Calendar throws an error when calling the non-existent
i18n.getLocalization method. With this patch, if Dojo is used with pure AMD
syntax dijit/Calendar will load the necessary CLDR data and work correctly.